Rename `pure-assert.js` to `assert.js`
